+++ brew/grains.py
@@ -282,6 +282,28 @@ class GrainAddition(object):
             u'dry_weight': round(self.get_dry_weight(), 2),
         }
 
+    @property
+    def gu(self):
+        return self.get_gravity_units()
+
+    def get_gravity_units(self):
+        """
+        Get the gravity units for the Grain Addition
+        :return: Gravity Units as PPG or HWE depending on units
+        :rtype: float
+
+        Gravity Units are really a property of whole grains, not of extracts.
+        Therefore the units will always be represented after the weight has
+        been converted from an extract weight to a cereal weight.
+        """
+        # Pick the attribute based on units
+        if self.units == IMPERIAL_UNITS:
+            attr = u'ppg'
+        if self.units == SI_UNITS:
+            attr = u'hwe'
+
+        return getattr(self.grain, attr) * self.get_cereal_weight()
+
     def to_dict(self):
         grain_data = self.grain.to_dict()
         return {u'name': grain_data.pop('name'),